First version of the web front end
The script will present the user with all of the scraped article, and
this is what the web front end is designed for. The html file will be
modified to include the html and css needed for displaying the newly
scraped article and thereby using these files as a template for how to
present the articles.