Actually use Jinja for rendering the feed
Previously the program would use python to manually assemble the HTML
for the feed page, but as of this update it will instead be using the
Jinja engine along with python lists to create the front page. Not only
is this safer, as this allows for proper escaping of the data coming
from the DB (which remember, originally came from outside sources and
were barely sanitized at all), but it also allows for easier
customisation, if lets say one would want a toggle to change whether the
links on the feed page actually lead to the source or maybe somewhere
else.