Add default value to limit input
Prior the limit input would not have a default value, and if the user
didn't input a value, this would mean that the script errored as the
query paramater "limit" would be given, but with the value of an empty
string, which would give error in the backend as the value would be
interpreted as a string (which is not allowed). Now it will instead use
a default value of 10.