Remove use of URLAndTitleList
Previously, the frontend feed page would use a special list created by
zipping the list of urls and the list of titles, to be able to loop
through the urls and the correnspondign title. This has been changed to
simplify the code, and now it will instead loop through the URL's and
title using numbers by taking the length of the url list.