Simplify the markArticleByID route
From now on, the id won't be send as a part of the URL, but instead in
the post body. This is done since the article id is specified at runtime
in the users browser using javascript variables, but jinja2 requires is
to be specified as a number at runtime ON THE SERVER for using the
url_for function to generate dynamic urls, so now it will instead send
this information in the body of the request.